+++ browser/CHANGELOG.md
@@ -10,6 +10,7 @@ This changelog covers all five packages, as they are (for now) updated as a whol
 - [#970](https://github.com/atomicdata-dev/atomic-server/issues/970) Add "show commit" button in History
 - [#968](https://github.com/atomicdata-dev/atomic-server/issues/968) Allow users to pick files by entering a subject into the file picker search bar.
 - [#969](https://github.com/atomicdata-dev/atomic-server/issues/969) Fix markdown editor sometimes doesn't update the value after saving.
+- [#975](https://github.com/atomicdata-dev/atomic-server/issues/975) Add create button to resource selector without classtype.
 - Fix markdown editor closing the edit page when a button is clicked.
 - Add an edit button to the default resource view.